+++ miho_buffered_rot_pytorch_gpu_further_more.py
@@ -301,7 +301,7 @@ def sampler4_par(n_par, m):
 
     n_par = n_par.repeat(m)
 
-    sidx = (torch.rand((m * nn, 4), device=device) * torch.stack((n_par, n_par-1, n_par-2, n_par-3)).permute(1,0)).type(torch.int)  
+    sidx = (torch.rand((m * nn, 4), device=device) * torch.stack((n_par, n_par-1, n_par-2, n_par-3)).permute(1,0)).type(torch.long)  
 
     for k in range(1,4):
         sidx[:, 0:k] = torch.sort(sidx[:, 0:k])[0]
@@ -356,7 +356,7 @@ def ransac_middle(pt1, pt2, dd, th_in=7, th_out=15, max_iter=500, min_iter=50, p
         undone_n = c_par_sz
         undone_i = torch.zeros(c_par_sz, device=device)
 
-        sidx_par = torch.zeros((c_par_sz, 4), dtype=torch.int, device=device)
+        sidx_par = torch.zeros((c_par_sz, 4), dtype=torch.long, device=device)
         good_sample_par = torch.zeros(c_par_sz, dtype=torch.bool, device=device)
 
         iter_count = 1
